% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms-timer.R
\name{hms_timer}
\alias{hms_timer}
\title{hms Timer}
\description{
A hms Timer is a \code{\link[hms:hms]{hms::hms()}} scalar which if
running has an attribute named start
that specifies the system time when the timer was started.
}
\details{
The elapsed time is the value of the scalar plus the difference
between the current system time and the system time when the timer was started.
}
\examples{
str(tmr_timer())
str(tmr_timer(1.5, start = TRUE))

x <- tmr_timer(1, start = TRUE)
print(x)
Sys.sleep(0.1)
print(x)
print(tmr_elapsed(x))
print(x)
}
