% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmstimer-package.R
\docType{package}
\name{hmstimer-package}
\alias{hmstimer}
\alias{hmstimer-package}
\title{hmstimer: 'hms' Based Timer}
\description{
Tracks elapsed clock time using a `hms::hms()` scalar, 
    which if running has an attribute named start 
    that specifies the system time when the timer was started.
    The elapsed time is the value of the scalar 
    plus the difference between the current system time 
    and the system time when the timer was started.
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
