% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-tmr.R
\name{local_timer}
\alias{local_timer}
\title{Local Timer}
\usage{
local_timer(title = TRUE, .local_envir = rlang::caller_env())
}
\arguments{
\item{title}{A flag indicating whether to print the timer with a title as the system call for where the timer was created or a string of the title to use.}

\item{.local_envir}{The environment to use for scoping}
}
\value{
xx
}
\description{
Local Timer
}
\examples{
local_timer()
}
