% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Default Parameter Descriptions}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{.local_envir}{The environment to use for scoping.}

\item{code}{Code to time.}

\item{digits}{A whole number of the number of decimal places.}

\item{print_title}{A flag specifying whether to print the title.}

\item{seconds}{A non-negative numeric scalar of the initial number of seconds.}

\item{srcref}{A flag specifying whether to print the source reference.}

\item{start}{A flag specifying whether to start the timer.}

\item{title}{A string of the title.}

\item{value}{A string of the title.}

\item{x}{A \code{\link[=hms_timer]{hms_timer()}}.}
}
\description{
Default parameter descriptions which may be overridden in individual
functions.
}
\keyword{internal}
