% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceiling.R
\name{tmr_ceiling}
\alias{tmr_ceiling}
\title{Ceiling hms Timer}
\usage{
tmr_ceiling(x)
}
\arguments{
\item{x}{A \code{\link[=hms_timer]{hms_timer()}}.}
}
\value{
A \code{\link[=hms_timer]{hms_timer()}}.
}
\description{
Rounds a \code{\link[=hms_timer]{hms_timer()}} up to the nearest second.
}
\examples{
tmr_ceiling(tmr_timer(18.9))
tmr_ceiling(tmr_timer(122.1))
}
\seealso{
Other round: 
\code{\link{tmr_floor}()},
\code{\link{tmr_format}()},
\code{\link{tmr_round}()}
}
\concept{round}
