% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floor.R
\name{tmr_floor}
\alias{tmr_floor}
\title{Floor hms Timer}
\usage{
tmr_floor(x)
}
\arguments{
\item{x}{A \code{\link[=hms_timer]{hms_timer()}}.}
}
\value{
A \code{\link[=hms_timer]{hms_timer()}}.
}
\description{
Rounds a \code{\link[=hms_timer]{hms_timer()}} down to the nearest second.
}
\examples{
tmr_floor(tmr_timer(18.9))
tmr_floor(tmr_timer(122.1))
}
\seealso{
Other round: 
\code{\link{tmr_ceiling}()},
\code{\link{tmr_format}()},
\code{\link{tmr_round}()}
}
\concept{round}
