% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{tmr_format}
\alias{tmr_format}
\title{Format hms Timer}
\usage{
tmr_format(x, digits = 3)
}
\arguments{
\item{x}{A hms scalar.}

\item{digits}{xx}
}
\value{
A character string.
}
\description{
Format hms Timer
}
\examples{
tmr_format(tmr_timer(61.36))
tmr_format(tmr_timer(61.36), digits = 4)
}
