% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-started.R
\name{tmr_is_started}
\alias{tmr_is_started}
\title{Is hms Timer Started}
\usage{
tmr_is_started(x)
}
\arguments{
\item{x}{A \code{\link[=hms_timer]{hms_timer()}}.}
}
\value{
A flag (TRUE or FALSE).
}
\description{
Tests if a \code{\link[=hms_timer]{hms_timer()}} is started (as indicated by the
presence of an attribute named start).
}
\examples{
tmr <- tmr_timer(start = TRUE)
print(tmr_is_started(tmr))
tmr <- tmr_stop(tmr)
print(tmr_is_started(tmr))
}
\seealso{
Other start_stop: 
\code{\link{tmr_elapsed}()},
\code{\link{tmr_is_stopped}()},
\code{\link{tmr_reset}()},
\code{\link{tmr_start}()},
\code{\link{tmr_stop}()},
\code{\link{tmr_timer}()}
}
\concept{start_stop}
