% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-stopped.R
\name{tmr_is_stopped}
\alias{tmr_is_stopped}
\title{Is hms Timer Stopped}
\usage{
tmr_is_stopped(x)
}
\arguments{
\item{x}{A \code{\link[=hms_timer]{hms_timer()}}.}
}
\value{
A flag.
}
\description{
Tests if a \code{\link[=hms_timer]{hms_timer()}} is stopped (as indicated by the
absence of an attribute named start).
}
\examples{
tmr <- tmr_timer(start = TRUE)
print(tmr_is_stopped(tmr))
tmr <- tmr_stop(tmr)
print(tmr_is_stopped(tmr))
}
\seealso{
Other start_stop: 
\code{\link{tmr_elapsed}()},
\code{\link{tmr_is_started}()},
\code{\link{tmr_print}()},
\code{\link{tmr_reset}()},
\code{\link{tmr_start}()},
\code{\link{tmr_stop}()},
\code{\link{tmr_timer}()}
}
\concept{start_stop}
