% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-titled.R
\name{tmr_is_titled}
\alias{tmr_is_titled}
\title{Is hms Timer Title}
\usage{
tmr_is_titled(x)
}
\arguments{
\item{x}{A \code{\link[=hms_timer]{hms_timer()}}.}
}
\value{
A flag (TRUE or FALSE).
}
\description{
Tests if a \code{\link[=hms_timer]{hms_timer()}} has a title (as indicated by the
presence of an attribute named start).
}
\examples{
tmr_is_titled(tmr_timer())
tmr_is_titled(tmr_timer(title = "my timer"))
}
