% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{tmr_print}
\alias{tmr_print}
\title{Print hms Timer}
\usage{
tmr_print(x, title = TRUE)
}
\arguments{
\item{x}{A \code{\link[=hms_timer]{hms_timer()}}.}

\item{title}{A flag indicating whether to print the timer with a title as the system call for where the timer was created or a string of the title to use.}
}
\value{
A character string.
}
\description{
Returns the elapsed time for a \code{\link[=hms_timer]{hms_timer()}} from the system time when the
timer was started and the current system time as an hms time.
}
\details{
The elapsed time is the value of the scalar plus the difference
between the current system time and the system time when the timer was started.
}
\examples{
x <- tmr_start(tmr_timer())
tmr_print(x)
}
\seealso{
Other print: 
\code{\link{tmr_title}()}
}
\concept{print}
