% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{tmr_reset}
\alias{tmr_reset}
\title{Reset hms Timer}
\usage{
tmr_reset(x, seconds = 0)
}
\arguments{
\item{x}{A \code{\link[=hms_timer]{hms_timer()}}.}

\item{seconds}{A non-negative numeric scalar of the initial number of seconds.}
}
\value{
A \code{\link[=hms_timer]{hms_timer()}}.
}
\description{
Resets a \code{\link[=hms_timer]{hms_timer()}} by creating a new one.
}
\examples{
tmr <- tmr_timer(10)
print(tmr)
tmr_reset(tmr)
}
\seealso{
Other start_stop: 
\code{\link{tmr_elapsed}()},
\code{\link{tmr_is_started}()},
\code{\link{tmr_is_stopped}()},
\code{\link{tmr_start}()},
\code{\link{tmr_stop}()},
\code{\link{tmr_timer}()}
}
\concept{start_stop}
