% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{tmr_round}
\alias{tmr_round}
\title{Round hms Timer}
\usage{
tmr_round(x, digits = 0)
}
\arguments{
\item{x}{A \code{\link[=hms_timer]{hms_timer()}}.}

\item{digits}{A count of the number of decimal places.}
}
\value{
A \code{\link[=hms_timer]{hms_timer()}}.
}
\description{
Rounds a \code{\link[=hms_timer]{hms_timer()}} after updating it to the elapsed time.
}
\details{
Negative values of digit are permitted.
}
\examples{
tmr_round(tmr_timer(18.9))
tmr_round(tmr_timer(18.9), 1)
tmr_round(tmr_timer(18.9), -1)
tmr_round(tmr_timer(121), -2) # 121 is rounded to 100 seconds
}
\seealso{
Other round: 
\code{\link{tmr_ceiling}()},
\code{\link{tmr_floor}()},
\code{\link{tmr_format}()}
}
\concept{round}
