% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{tmr_title<-}
\alias{tmr_title<-}
\title{Set Title hms Timer}
\usage{
tmr_title(x) <- value
}
\arguments{
\item{x}{A \code{\link[=hms_timer]{hms_timer()}}.}

\item{value}{A string of the title.}
}
\value{
A copy of the \code{\link[=hms_timer]{hms_timer()}} with the new title.
}
\description{
Sets the title of a \code{\link[=hms_timer]{hms_timer()}}.
}
\examples{
tmr <- tmr_timer(title = "A title")
tmr_print(tmr)
tmr_title(tmr) <- "A different title"
tmr_print(tmr)
tmr_title(tmr) <- NULL
tmr_print(tmr)
}
\seealso{
\code{\link[=tmr_title]{tmr_title()}}
}
