% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{tmr_title}
\alias{tmr_title}
\title{Title hms Timer}
\usage{
tmr_title(x)
}
\arguments{
\item{x}{A \code{\link[=hms_timer]{hms_timer()}}.}
}
\value{
A string of the hms timer title or NULL if undefined.
}
\description{
Title hms Timer
}
\examples{
tmr_title(tmr_start(tmr_timer()))
tmr_title(tmr_start(tmr_timer(title = "new one")))
tmr_title(tmr_start(tmr_timer(title = FALSE)))
}
\seealso{
Other print: 
\code{\link{tmr_print}()}
}
\concept{print}
