% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with-timer.R
\name{with_timer}
\alias{with_timer}
\title{With Timer}
\usage{
with_timer(code)
}
\arguments{
\item{code}{Code to time.}
}
\value{
The result of executing code.
}
\description{
With Timer
}
\examples{
with_timer(Sys.sleep(0.1))

with_timer({
 for(i in 1:2) {
  Sys.sleep(0.1)
 }
})
}
