% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmstimer-package.R
\docType{package}
\name{hmstimer-package}
\alias{hmstimer}
\alias{hmstimer-package}
\title{hmstimer: 'hms' Based Timer}
\description{
Tracks elapsed clock time using a 'hms()' scalar 
    (inherits from 'difftime()' with seconds as the unit).
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
