% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmstimer-package.R
\docType{package}
\name{hmstimer-package}
\alias{hmstimer}
\alias{hmstimer-package}
\title{hmstimer: 'hms' Based Timer}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tracks elapsed clock time using a `hms::hms()` scalar. It was was originally developed to time Bayesian model runs. It should not be used to estimate how long extremely fast code takes to execute as the package code adds a small time cost.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/hmstimer}
  \item \url{https://poissonconsulting.github.io/hmstimer/}
  \item Report bugs at \url{https://github.com/poissonconsulting/hmstimer/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})
}

Other contributors:
\itemize{
  \item Nadine Hussein \email{nadine@poissonconsulting.ca} (\href{https://orcid.org/0000-0003-4470-8361}{ORCID}) [contributor]
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
