% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-timer.R
\name{local_timer}
\alias{local_timer}
\title{Local Timer}
\usage{
local_timer(.local_envir = rlang::caller_env())
}
\arguments{
\item{.local_envir}{The environment to use for scoping}
}
\value{
An invisible copy of the hmstimer object.
}
\description{
Local Timer
}
\examples{
fun <- function() {
  local_timer()
}
fun()
}
