% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-timer.R
\name{local_timer}
\alias{local_timer}
\title{Local Timer}
\usage{
local_timer(..., title = "", srcref = TRUE, .local_envir = rlang::caller_env())
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{title}{A string of the title.}

\item{srcref}{A flag specifying whether to print the source reference.}

\item{.local_envir}{The environment to use for scoping.}
}
\description{
Called for the side effect of providing a message of the time
required to execute the rest of the function.
}
\examples{
fun <- function() {
  local_timer()
  Sys.sleep(0.1)
  10
}
fun()
}
\seealso{
\code{\link[=with_timer]{with_timer()}}
}
