% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{#' Parameter Descriptions for hmstimer Functions}
\arguments{
\item{x}{A \code{\link[=hms_timer]{hms_timer()}}.}

\item{seconds}{A non-negative numeric scalar of the initial number of seconds.}

\item{start}{A flag indicating whether to start the timer.}

\item{title}{A flag indicating whether to print the timer with a title as the system call for where the timer was created or a string of the title to use.}

\item{digits}{A whole number of the number of decimal places.}
}
\description{
#' Parameter Descriptions for hmstimer Functions
}
\keyword{internal}
