% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\alias{parameters}
\alias{arguments}
\alias{args}
\title{Parameter Descriptions}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{code}{A line or block of R code.}

\item{digits}{A count of the number of decimal places.}

\item{print_title}{A flag specifying whether to print the title.}

\item{seconds}{A non-negative numeric scalar of the initial number of seconds.}

\item{srcref}{A flag specifying whether to print the source reference.}

\item{start}{A flag specifying whether to start the timer.}

\item{title}{A string of the title.}

\item{value}{A string of the title.}

\item{x}{A \code{\link[=hms_timer]{hms_timer()}}.}

\item{.local_envir}{The environment to use for scoping.}
}
\description{
Default parameter descriptions which may be overridden in individual
functions.
}
\details{
A flag is a non-missing logical scalar.

A string is a non-missing character scalar.

A count is a non-missing non-negative integer scalar or double
scalar with no fractional part.
}
\keyword{internal}
