% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{tmr_format}
\alias{tmr_format}
\title{Format hms Timer}
\usage{
tmr_format(x, digits = 3, title = FALSE)
}
\arguments{
\item{x}{A \code{\link[=hms_timer]{hms_timer()}}.}

\item{digits}{A whole number of the number of decimal places.}

\item{title}{A flag indicating whether to print the timer with a title as the system call for where the timer was created or a string of the title to use.}
}
\value{
A character string.
}
\description{
Converts a \code{\link[=hms_timer]{hms_timer()}} to a string of the clock time
after rounding it to the number of digits.
}
\details{
Negative values of digit are not permitted.
}
\examples{
tmr_format(tmr_timer(61.66))
tmr_format(tmr_timer(61.66), digits = 0)
}
\seealso{
Other round: 
\code{\link{tmr_ceiling}()},
\code{\link{tmr_floor}()},
\code{\link{tmr_round}()}
}
\concept{round}
