% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{tmr_timer}
\alias{tmr_timer}
\title{Create hms Timer}
\usage{
tmr_timer(seconds = 0, start = FALSE)
}
\arguments{
\item{seconds}{A non-negative numeric scalar of the initial number of seconds.}

\item{start}{A flag indicating whether to start the timer.}
}
\value{
A \code{\link[=hms_timer]{hms_timer()}}.
}
\description{
Creates a \code{\link[=hms_timer]{hms_timer()}}.
}
\examples{
tmr <- tmr_timer()
print(tmr)
class(tmr)
}
\seealso{
Other start_stop: 
\code{\link{tmr_elapsed}()},
\code{\link{tmr_is_started}()},
\code{\link{tmr_is_stopped}()},
\code{\link{tmr_print}()},
\code{\link{tmr_reset}()},
\code{\link{tmr_start}()},
\code{\link{tmr_stop}()}
}
\concept{start_stop}
