% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{tmr_title}
\alias{tmr_title}
\title{Get Title hms Timer}
\usage{
tmr_title(x)
}
\arguments{
\item{x}{A \code{\link[=hms_timer]{hms_timer()}}.}
}
\value{
A flag of the title.
}
\description{
Returns a flag (character vector) of the title.
}
\examples{
tmr_title(tmr_timer())
tmr_title(tmr_timer(title = ""))
tmr_title(tmr_timer(title = "A Title"))
}
\seealso{
\code{\link[=tmr_title<-]{tmr_title<-()}}
}
