% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with-timer.R
\name{with_timer}
\alias{with_timer}
\title{With Timer}
\usage{
with_timer(code, ..., srcref = FALSE)
}
\arguments{
\item{code}{Code to time.}

\item{...}{Must not be used.}

\item{srcref}{A flag specifying whether to print the source reference.}
}
\value{
The result of executing the code.
}
\description{
With Timer
}
\examples{
fun <- function() {
  Sys.sleep(0.1)
  10
}
with_timer(fun())

with_timer({
  for (i in 1:2) {
    Sys.sleep(0.1)
  }
  20
})
}
\seealso{
\code{\link[=local_timer]{local_timer()}}
}
